+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 - Will attempt to extract a changelog.
+- Added `built`, `changelog-entry`, and `tag-version` outputs.
 
 # 0.1.0